added calendar function

+++ _server/core/views.py
@@ -1,11 +1,13 @@
+from datetime import datetime
 from django.shortcuts import render
 from django.conf  import settings
 from django.shortcuts import redirect
 from django.http import HttpResponse, HttpResponseBadRequest, JsonResponse
-from .models import Lesson
+from .models import Lesson, StudioEvent
 import json
 import os
 from django.contrib.auth.decorators import login_required
+from django.views.decorators.csrf import csrf_exempt
 
 # Load manifest when server launches
 MANIFEST = {}
@@ -28,7 +30,10 @@ def index(req):
 @login_required
 def book_lesson(req):
     if  not req.method == "POST":
-        return redirect("/not-found")
+        response = JsonResponse({ "success": False, "message": "Invalid request method" })
+        response.status_code = 404
+        return response
+
     data = json.loads(req.body)
     print(data)
     if not data["startTime"] or not data["date"] or not data["duration"]:
@@ -57,20 +62,28 @@ def cancel_lesson(req, lesson_id):
     if not req.method == "DELETE":
         response = JsonResponse({ "success": False, "message": "Invalid request method" })
         response.status_code = 400
+        return response
         
     if not req.user.is_authenticated:
         response = JsonResponse({ "success": False, "message": "User not authenticated" })
         response.status_code = 401
         return response
     try:
-        print(lesson_id)
         lesson = Lesson.objects.get(id=lesson_id)
+        lesson_day_of_week = str(lesson.date)
+        current_day_of_week = datetime.now().strftime("%A")
+        lesson_hour_in_military = int(lesson.startTime.split(":")[0]) + 12 if "pm" in lesson.startTime else int(lesson.startTime.split(":")[0])
+        current_hour_in_military = int(datetime.now().strftime("%H"))
+        if lesson_day_of_week == current_day_of_week and lesson_hour_in_military - current_hour_in_military < 2 and lesson_hour_in_military - current_hour_in_military > 0:
+            return HttpResponseBadRequest("Cannot cancel a lesson within 2 hours of the start time")
+
         lesson.delete()
     except Lesson.DoesNotExist:
         return HttpResponseBadRequest("Lesson not found")
     response = HttpResponse("Lesson cancelled successfully")
     response.status_code = 200
     return response
+
 
 @login_required
 def get_account(req):
@@ -94,10 +107,13 @@ def get_account(req):
 
     return JsonResponse({ "success": True ,"user": user})
 
+
 @login_required
 def get_lessons(req):
     if not req.method == "GET":
-        return redirect("/not-found")
+        response = JsonResponse({ "success": False, "message": "Invalid request method" })
+        response.status_code = 404
+        return response
     
     if not req.user.is_authenticated:
         return redirect("/login")
@@ -114,4 +130,44 @@ def get_lessons(req):
             "price": lesson.price
         })
     return JsonResponse(lessons_list, safe=False)
+
+
+@csrf_exempt
+def add_studio_event(req):
+    if not req.method == "POST":
+        response = JsonResponse({ "success": False, "message": "Invalid request method" })
+        response.status_code = 404
+        return response
+    data = json.loads(req.body)
+    print(data)
+    if not data["title"] or not data["start"]:
+        return HttpResponseBadRequest("Missing required fields")
+    studio_event = StudioEvent(name = data["title"],
+        start = data["start"],
+        end = data["end"],
+        allDay = data["allDay"]
+    )
+    studio_event.save()
+    response = HttpResponse("Studio event added successfully")
+    response.status_code = 201
+    return response
+
+
+def get_studio_events(req):
+    if not req.method == "GET":
+        response = JsonResponse({ "success": False, "message": "Invalid request method" })
+        response.status_code = 404
+        return response
+    
+    studio_events = StudioEvent.objects.all()
+    studio_events_list = []
+    for studio_event in studio_events:
+        studio_events_list.append({
+            "id": studio_event.id,
+            "title": studio_event.name,
+            "start": studio_event.start,
+            "end": studio_event.end,
+            "allDay": studio_event.allDay
+        })
+    return JsonResponse(studio_events_list, safe=False)
